Fixed bug in depends.mk
If package-common.mk was included in a Makefile that also included depends.mk,
package-common.mk would get caught by the package-% rule, which would cause the
target to fail. Added a package-%.mk phony target that does nothing to prevent
this from happening and prevent it from reoccurring in the future with
different Makefiles.