busybox 1.13.1 build 2
Fixed a dangerous SSL/TLS bug in wget. Prior to this, wget would not
verify x509 certificates when connecting via HTTPS, even when the
--no-check-certificate flag was not specified; instead it would connect
without issue or warning, even if the certificate was invalid.

This has been fixed in this build. The --no-check-certificate flag can
still be used if you need to connect over HTTPS without checking the
certificate.