busybox 1.31.1 build 3
Added a shim for bash so that if the real Bash is installed it will be
used, but if not BusyBox's ash will be used. Many applications
(including some in the base system) require the presence of /bin/bash
to work, but some applications in pkgsrc work only if /bin/bash is
pointing to the real Bash shell. This satisifes both cases. Also check
in /usr/local in case Bash is installed there.