Added arch check to linux buildsystem
Previously, the linux buildsystem would reuse the existing kernel source
tree without cleaning if the source tree's kernel version matched the
package's kernel version. With the introduction, this creates a problem
if the same /usr/src tree is used to compile a kernel for one
architecture and then a kernel for another architecture: the kernel
image produced for the second architecture will contain a mix of
architectures, rendering it inoperable.

To fix this, the check has been expanded to check that both the kernel
versions and target architectures match. If not, a new kernel source
tree is created.